Merge Stack child options only for current child of stack
Currently child options are applied only for visible children. During push animations, both children are visible.
If mergeOptions is called on the disappearing child during push animation, wrong options will be applied.

This commit changes the conditional so that child options are merged to stack if the child is also the current child.