Fix Topbar custom component flicker across screens #3864 (#5110)
When a topBar.backgroundComponent is defined, check if a corresponding ReactView has already been created with matching componentName and componentId. If such view is found, reuse it instead of creating another view.

This change will fix a flickering in the TopBar background component when a screen was pushed into a stack with a background component.